pyproject.toml
[project]
name = "livraria"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "santiago schons", email = ""},
]
dependencies = [
    "django>=5.0.3",
    "djangorestframework>=3.15.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
